adds basic refreshr cards and logic
